backend/users/urls.py
from django.urls import path
from rest_framework_simplejwt.views import (
    TokenObtainPairView,
    TokenRefreshView,
)
from .views import RegisterView, ProfileView, AdminUserListView, AdminUserExportView

urlpatterns = [
    path('register/', RegisterView.as_view(), name='register'),
    path('profile/', ProfileView.as_view(), name='profile'),
    path('token/', TokenObtainPairView.as_view(), name='token_obtain_pair'),
    path('token/refresh/', TokenRefreshView.as_view(), name='token_refresh'),

    # Admin Endpoints
    path('admin/users/', AdminUserListView.as_view(), name='admin-user-list'),
    path('admin/export/', AdminUserExportView.as_view(), name='admin-user-export'),
]
