backend/requirements.txt
Django>=4.2,<5.0
djangorestframework
djangorestframework-simplejwt
drf-spectacular
django-cors-headers
django-filter
psycopg2-binary
python-dotenv
Pillow
gunicorn
dj-database-url
whitenoise
